Only link fixcommitter if it doesn't exist

+++ setup-misc.sh
@@ -6,6 +6,6 @@ set -e
 TARGET=/usr/local/bin
 SOURCE=$HOME/workspace/capi-release/scripts # notice this NOT capi-workspace, if you are adding content to this file (REFACTOR!)
 
-if [ ! -f $TARGET/fixcommitter ] ; then
+if [ ! -h $TARGET/fixcommitter ] && [ ! -f $TARGET/fixcommitter ]; then
 	ln -s $SOURCE/fixcommitter $TARGET/fixcommitter
 fi
